FLIZ-452 feat: /notification/reservation 페이지 네트워크 최적화
